requirements.txt
Flask
Flask-Login
Flask-Migrate
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
gunicorn
werkzeug
pandas
openpyxl
celery
redis